+++ src/Type/Payment/TransactionPartnerUser.php
@@ -19,6 +19,8 @@ final readonly class TransactionPartnerUser implements TransactionPartner
         #[ArrayMap(PaidMediaValue::class)]
         public ?array $paidMedia = null,
         public ?string $paidMediaPayload = null,
+        public ?int $subscriptionPeriod = null,
+        public ?string $gift = null,
     ) {}
 
     public function getType(): string